feat: 회사 MyPage 분리
